requirements.txt
certifi==2020.6.20
cffi==1.14.2
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.2.0
networkx==2.4
numpy==1.19.1
Pillow==7.2.0
pkg-resources==0.0.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.8.13
tqdm==4.48.2
webp==0.1.0a15
